requirements.txt
Click==7.0
Flask==1.0.3
googletrans==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
pony==0.7.10
psycopg2==2.8.2
python-dotenv==0.10.3
Werkzeug==0.15.4
